refactor(pagination): split into smaller files for readability
