+++ .travis.yml
@@ -27,18 +27,18 @@ matrix:
     - cmake --version
     - cargo coverage || true
   - env: RUSTFMT
-    rust: 1.31.0  # `stable`: Locking down for consistent behavior
+    rust: 1.33.0  # `stable`: Locking down for consistent behavior
     install:
     - rustup component add rustfmt
     script:
     - cargo fmt -- --check
   - env: RUSTFLAGS="-D warnings"
-    rust: 1.31.0  # `stable`: Locking down for consistent behavior
+    rust: 1.33.0  # `stable`: Locking down for consistent behavior
     install:
     script:
     - cargo check --tests
   - env: CLIPPY
-    rust: 1.31.0  # `stable`: Locking down for consistent behavior
+    rust: 1.33.0  # `stable`: Locking down for consistent behavior
     install:
       - rustup component add clippy
     script:
